refactor(ai-provider): align provider module layout
