+++ src/ai/provider/custom/aihubmix/aihubmixProvider.ts
@@ -12,7 +12,7 @@ import { loadApiKey, withoutTrailingSlash } from '@ai-sdk/provider-utils';
 
 import type { Model } from '@/data/types/model';
 
-import { isOpenAIChatCompletionOnlyModel, isOpenAILLMModel } from '../../utils/model';
+import { isOpenAIChatCompletionOnlyModel, isOpenAILLMModel } from '../../../utils/model';
 
 export const AIHUBMIX_PROVIDER_NAME = 'aihubmix' as const;
 const APP_CODE_HEADER = { 'APP-Code': 'MLTG2087' };
